refactor: 型定義ファイルはComponentからも呼ばれていたのでディレクトリを移動した
